requirements.txt
openpyxl
requests
openai
pillow 